src/main/java/com/github/smkozh/restaurantvoting/web/restaurant/RestaurantRestController.java
package com.github.smkozh.restaurantvoting.web.restaurant;

import com.github.smkozh.restaurantvoting.model.Restaurant;
import org.springframework.http.MediaType;
import org.springframework.transaction.annotation.Transactional;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping(value = RestaurantRestController.REST_URL, produces = MediaType.APPLICATION_JSON_VALUE)
public class RestaurantRestController extends AbstractRestaurantController {
    static final String REST_URL = "/api/restaurants";

    @Transactional
    @GetMapping("/with-menus")
    public List<Restaurant> getAllWithTodayMenus() {
        return super.getAllWithTodayMenus();
    }
}
