src/test/java/com/github/smkozh/restaurantvoting/web/restaurant/RestaurantRestControllerTest.java
package com.github.smkozh.restaurantvoting.web.restaurant;

import com.github.smkozh.restaurantvoting.web.AbstractControllerTest;
import org.junit.jupiter.api.Test;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithUserDetails;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import static com.github.smkozh.restaurantvoting.web.restaurant.RestaurantTestData.*;
import static com.github.smkozh.restaurantvoting.web.user.UserTestData.USER_MAIL;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

class RestaurantRestControllerTest extends AbstractControllerTest {

    private static final String REST_URL = RestaurantRestController.REST_URL + "/";

    @Test
    @WithUserDetails(value = USER_MAIL)
    void getAllWithTodayMenus() throws Exception {
        perform(MockMvcRequestBuilders.get(REST_URL + "with-menus"))
                .andExpect(status().isOk())
                .andDo(print())
                .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
                .andExpect(RESTAURANT_WITH_MENU_MATCHER.contentJson(restaurant1, restaurant2WithTodayMenu, restaurant3));
    }
}